Move notification to user profile
