+++ frontend/src/modules/signup-old/signup-form.jsx
@@ -20,6 +20,7 @@ import { Trans, t } from '@lingui/macro'
 import { useLingui } from '@lingui/react'
 import CatTagSelect from '../../components/cat-tag-select/cat-tag-select'
 import FormItem from 'antd/lib/form/FormItem'
+import { useRouter } from 'next/router'
 
 const { sectorOptions } = UIStore.currentState
 
@@ -151,7 +152,7 @@ const SignupForm = ({
     organisationType: s.organisationType,
     nonMemberOrganisations: s.nonMemberOrganisations,
   }))
-
+  const router = useRouter()
   const [noOrg, setNoOrg] = useState(false)
   const [formInitialValues, setInitialValues] = useState(false)
   const [pubEmail, setPubEmail] = useState({
@@ -165,6 +166,7 @@ const SignupForm = ({
     orgValue: false,
   })
 
+  const [enableNotification, setEnableNotification] = useState(false)
   const defaultFormSchema = useDefaultFormSchema()
 
   const formRef = useRef()
@@ -240,6 +242,7 @@ const SignupForm = ({
   useEffect(() => {
     if (initialValues) {
       handleChangePublicEmail(initialValues.publicEmail)
+      setEnableNotification(initialValues.chatEmailNotifications)
     }
     if (initialValues && initialValues?.org === null) {
       handleChangePrivateCitizen({ target: { checked: true } })
@@ -304,6 +307,14 @@ const SignupForm = ({
   }
 
   const required = (value) => (value ? undefined : 'Required')
+
+  const handleNotification = (value) => {
+    setEnableNotification(value)
+    router.push({
+      pathname: '/unsubscribe-chat',
+      query: { id: initialValues.id },
+    })
+  }
 
   return (
     <Form layout="vertical">
@@ -539,6 +550,16 @@ const SignupForm = ({
                       }}
                     </Field>
                   </Form.Item>
+                  <div className="notification-container">
+                    <Switch
+                      key="enableNotification"
+                      name="enableNotification"
+                      size="small"
+                      checked={enableNotification}
+                      onChange={handleNotification}
+                    />
+                    Receive Email Notifications for Forum Channel Activity
+                  </div>
                 </div>
               </div>
             </>
