+++ frontend/src/modules/signup-old/styles.module.scss
@@ -76,6 +76,13 @@
   justify-content: space-between;
 
   :global {
+    .notification-container {
+      display: flex;
+      gap: 10px;
+      button {
+        margin: 3px 0px;
+      }
+    }
     h2 {
       font-size: 16px;
       line-height: 22px;
