+++ frontend/src/modules/stakeholder-detail/view.jsx
@@ -228,7 +228,6 @@ const StakeholderDetail = ({
   const [ownedResourcesPage, setOwnedResourcesPage] = useState(0)
   const [bookedResourcesPage, setBookedResourcesPage] = useState(0)
   const [warningVisible, setWarningVisible] = useState(false)
-  const [enableNotification, setEnableNotification] = useState(false)
 
   const prevValue = usePrevious(data)
   const { id } = router.query
@@ -311,7 +310,6 @@ const StakeholderDetail = ({
           .get(`/detail/stakeholder/${id}`)
           .then((d) => {
             setData(d.data)
-            setEnableNotification(d.data.chatEmailNotifications)
             getOwnedResources(0)
             getBookedResources(0)
           })
@@ -368,14 +366,6 @@ const StakeholderDetail = ({
         }
       })
     }
-  }
-
-  const handleNotification = (value) => {
-    setEnableNotification(value)
-    router.push({
-      pathname: '/unsubscribe-chat',
-      query: { id: data.id },
-    })
   }
 
   if (!data) {
@@ -532,16 +522,6 @@ const StakeholderDetail = ({
                       </List.Item>
                     )}
                   </List>
-                  <div className="notification-container">
-                    <Switch
-                      key="enableNotification"
-                      name="enableNotification"
-                      size="small"
-                      checked={enableNotification}
-                      onChange={handleNotification}
-                    />
-                    Receive Email Notifications for Forum Channel Activity
-                  </div>
                 </div>
               </CardComponent>
               <CardComponent title={<Trans>Contact info</Trans>}>
